small architecture refactor, orders end-points in progress!
