+++ src/api/routers.py
@@ -1,6 +1,6 @@
-from src.api.users import router as users_router
-from src.api.cars import router as cars_router
-from src.api.orders import router as orders_router
+from src.api.routes.users import router as users_router
+from src.api.routes.cars import router as cars_router
+from src.api.routes.orders import router as orders_router
 
 
 all_routers = [
